chore: Update dependencies and add Gunicorn for development server
This commit updates the `requirements.txt` file to include the latest versions of the dependencies. It also adds `gunicorn==22.0.0` to the list of dependencies, which will be used for running the development server.

The purpose of this change is to ensure that the project is using the latest stable versions of the dependencies and to provide a more robust and performant development server with Gunicorn.