chore: Update n_workers calculation in main.py
This commit updates the calculation of `n_workers` in the `main.py` file. The previous calculation used the formula `(multiprocessing.cpu_count() * 2) + 1,` which resulted in an extra comma at the end. This commit removes the extra comma to fix the syntax error.

The purpose of this change is to ensure that the `n_workers` variable is calculated correctly and to fix the syntax error in the code.