chore: Update pi-lotto package.json and public/index.html
This commit updates the pi-lotto package.json file to include the "ws" dependency with version "^8.17.0". Additionally, it updates the public/index.html file to add the "https://sdk.minepi.com/pi-sdk.js" script tag. These changes ensure that the project has the necessary dependencies and includes the required script for the Pi SDK.