chore: Add database triggers for Ticket model
This commit adds database triggers for the Ticket model. The `after_insert_ticket` function is called after a new ticket is inserted into the database, creating a new LottoStats entry for the ticket. The `after_update_ticket` function is called after a ticket is updated, updating the corresponding LottoStats entry with the updated numbers played. This will ensure that the LottoStats table stays synchronized with the Ticket table.