chore: Add create_account_transaction function
This commit adds the create_account_transaction function to the transactions.py file in the utils directory. This function is responsible for creating an account transaction for a user based on the transaction type, amount, and reference ID provided. It also updates the user's balance based on the transaction type.